refactor: simplify import logic in telemetry and tools utilities
